requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
Django==3.2.15
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.3.0
